redirect to dashboard if already loggedin
