+++ management/app/providers.tsx
@@ -3,7 +3,7 @@
 import { SessionProvider } from "next-auth/react";
 
 type Props = {
-  children?: React.ReactNode;
+  children: React.ReactNode
 };
 
 export default function NextAuthProvider({ children }: Props) {
@@ -12,8 +12,4 @@ export default function NextAuthProvider({ children }: Props) {
         {children}
     </SessionProvider>
   )
-};
-
-NextAuthProvider.defaultProps = {
-  children: null,
 };
